Updated database
Added new column to Post called votes which contains the votes of the post.
Added trigger to automatically update the vote on a post if the user has already voted on it.
Added trigger to automatically update the votes column of a post at each insert in the table Vote.